+++ src/briefy/ws/views/devops/lb_heartbeat.py
@@ -2,6 +2,7 @@
 
 ref: https://github.com/mozilla-services/cliquet/blob/master/cliquet/views/heartbeat.py
 """
+from pyramid.request import Request
 from pyramid.security import NO_PERMISSION_REQUIRED
 
 from cornice import Service
@@ -14,7 +15,7 @@ lbheartbeat = Service(
 
 
 @lbheartbeat.get(permission=NO_PERMISSION_REQUIRED)
-def get_lbheartbeat(request):
+def get_lbheartbeat(request: Request) -> dict:
     """Return successful healthy response.
 
     If the load-balancer tries to access this URL and fails, this means the