+++ src/briefy/ws/auth/__init__.py
@@ -5,6 +5,7 @@ from pyramid.httpexceptions import HTTPUnauthorized as BaseHTTPUnauthorized
 from webob import Response
 
 import json
+import newrelic.agent
 
 
 class AuthenticatedUser(BaseUser):
@@ -37,6 +38,10 @@ def user_factory(request) -> AuthenticatedUser:
     user_id = request.authenticated_userid
     if user_id:
         data = request.jwt_claims
+        # add all user data to the newrelic custom attributes
+        newrelic.agent.add_custom_parameter('user_id', user_id)
+        for key, value in data.items():
+            newrelic.agent.add_custom_parameter(key, str(value))
         return AuthenticatedUser(user_id, data)
     else:
         return None