+++ HISTORY.rst
@@ -8,6 +8,7 @@ History
     * BODY-64: Add like filter. (ericof)
     * BODY-88: Improve apply_secutiry to use context permissions (rudaporto).
     * BODY-89: Improve logging. (ericof).
+    * BODY-90: Remove leading underscore from node names. (ericof).
 
 1.1.1 (2016-10-04)
 ------------------